Moved path resolution from config to an extension trait since we'll need it for images
